+++ lib/ContextStackTrait.php
@@ -15,7 +15,7 @@ namespace Sabre\Xml;
  * override its settings for document fragments, and easily restore it again
  * later.
  *
- * @copyright Copyright (C) 2013-2014 fruux GmbH. All rights reserved.
+ * @copyright Copyright (C) 2009-2015 fruux GmbH (https://fruux.com/).
  * @author Evert Pot (http://evertpot.com/)
  * @license http://sabre.io/license/ Modified BSD License
  */