+++ lib/Element.php
@@ -10,36 +10,13 @@ use XMLReader;
  * Elements are responsible for serializing and deserializing part of an XML
  * document into PHP values.
  *
- * It's a superset of XmlSerializable. It just adds a static method that allows
- * values to be deserialized as well.
+ * It combines XmlSerializable and XmlDeserializable into one logical class
+ * that does both.
  *
- * @copyright Copyright (C) 2013-2014 fruux GmbH (https://fruux.com/).
+ * @copyright Copyright (C) 2009-2015 fruux GmbH (https://fruux.com/).
  * @author Evert Pot (http://evertpot.com/)
  * @license http://code.google.com/p/sabredav/wiki/License Modified BSD License
  */
-interface Element extends XmlSerializable {
-
-    /**
-     * The deserialize method is called during xml parsing.
-     *
-     * This method is called statictly, this is because in theory this method
-     * may be used as a type of constructor, or factory method.
-     *
-     * Often you want to return an instance of the current class, but you are
-     * free to return other data as well.
-     *
-     * Important note 2: You are responsible for advancing the reader to the
-     * next element. Not doing anything will result in a never-ending loop.
-     *
-     * If you just want to skip parsing for this element altogether, you can
-     * just call $reader->next();
-     *
-     * $reader->parseInnerTree() will parse the entire sub-tree, and advance to
-     * the next element.
-     *
-     * @param Reader $reader
-     * @return mixed
-     */
-    static function xmlDeserialize(Reader $reader);
+interface Element extends XmlSerializable, XmlDeserializable {
 
 }