+++ tests/Sabre/XML/ContextStackTest.php
@@ -5,7 +5,7 @@ namespace Sabre\Xml;
 /**
  * Test for the ContextStackTrait
  *
- * @copyright Copyright (C) 2007-2014 fruux GmbH. All rights reserved.
+ * @copyright Copyright (C) 2009-2015 fruux GmbH (https://fruux.com/).
  * @author Evert Pot (http://evertpot.com/)
  * @license http://sabre.io/license/ Modified BSD License
  */